+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Next
 ====
 
 - File path is now passed to template as ``filepath`` in the ``sync native`` command.
+- CLI can now invite users to Preset from a YAML file created by ``export-users``.
 
 Version 0.1.0 - 2022-09-09
 ==========================